+++ CHANGELOG.md
@@ -10,6 +10,7 @@ See [keep a changelog] for information about writing changes to this log.
 
 ### Added
 
+- Added mail service and templates for mails.
 - Added SMS gateway and command to send test SMS.
 - Added API client and commands to test it.
 - Metrics bundlen added.